skiplevel-agent/pyproject.toml
[project]
name = "skiplevel-agent"
version = "0.1.0"
description = "An AI-powered growth system for engineers"
authors = [
    {name = "atolat", email = "your-email@example.com"}
]
dependencies = [
    "langchain==0.3.23",
    "langgraph==0.3.31",
    "chainlit==2.3.0",
    "qdrant-client==1.13.3",
    "openai==1.75.0",
    "python-dotenv==1.1.0",
    "pydantic==2.10.0",
    "langsmith==0.3.32",
    "langchain-openai==0.3.14",
    "grandalf==0.7"
]
requires-python = ">=3.9"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 100
target-version = "py39"

[tool.hatch.build.targets.wheel]
packages = ["skiplevel_agent"] 